// packages/core/src/javaStreamables.d.ts
export interface MMREntryNumber {
    scale: number;
    unscaled: bigint;
}
// ...
export interface MMRData {
    data: Uint8Array;
    value: bigint;
}
// ...
export interface MMREntry {
    row: number;
    entryNumber: MMREntryNumber;
    mmrData: MMRData;
}